put persistence into a stores folder, consisting of a protocol and its implementations
